infra/pyproject.toml
[project]
name = "laravel-app"
version = "0.0.1"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.213.0",
    "constructs>=10.4.2",
    "dotenv>=0.9.9",
]
